pyproject.toml
[tool.poetry]
name = "pyfireconsole"
version = "0.0.2"
description = ""
authors = ["Makoto Tanji <tanji.makoto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0"
google-cloud-firestore = "^2.11.1"
google-auth = "^2.22.0"
inflect = "^7.0.0"
pydantic = "^2.1.1"
ipython = "^8.14.0"


[tool.poetry.group.dev.dependencies]
wheel = "^0.41.0"
twine = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output=2
line_length=180
